% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_signif.R
\name{cb_signif}
\alias{cb_signif}
\title{Determining number of lags in the cross-basis model}
\usage{
cb_signif(data, argvar = "lin", arglag = "lin", knots = NULL, startlag = 16)
}
\arguments{
\item{data}{Dataset.}

\item{argvar}{Shape of relationship between endpoint exposure and outcome.
Value entered for \code{argvar} in the cross-basis formula.}

\item{arglag}{Shape of relationship between lagged exposure and outcome.
Value entered for \code{arglag} in the cross-basis formula.}

\item{knots}{Degrees of freedom for the \code{arglag} argument, if any.}

\item{startlag}{Maximum number of lags to include in the model.}
}
\description{
This function determines the number of lags that should be used in the
cross-basis model. It begins with the maximum number of interest
(the startlag option). Each time a cross-basis is created, the final lag
is evaluated using unadjusted and adjusted models to determine significance
at alpha = 0.05 and alpha = 0.10. If the final lag is not significant at one
of these alphas, the model is run again with the oldest lag removed.

The output includes the settings for the final cross-basis selected,
with the final lag being the oldest with significance at alpha = 0.05 if
applicable or, if none have significance at alpha = 0.05, the oldest final
lag with significance at alpha = 0.10 is retained. If no final lags have
significance at alpha = 0.05 or 0.10, the smallest possible cross-basis is
retained, with the number of lags being 1 plus the number of degrees of
freedom. For linear models, one lag was retained.
}
