% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_signif_green.R
\name{cb_signif_green}
\alias{cb_signif_green}
\title{Determining number of lags in the cross-basis model with green space}
\usage{
cb_signif_green(
  data,
  argvar = "lin",
  arglag1 = "lin",
  knots1 = NULL,
  arglag2 = NULL,
  knots2 = NULL,
  startlag = 16,
  greens = c("gt", "gw")
)
}
\arguments{
\item{data}{Dataset.}

\item{argvar}{Shape of relationship between endpoint exposure and outcome.
Value entered for \code{argvar} in the cross-basis formula.}

\item{arglag1}{Shape of relationship between lagged exposure and outcome.
Value entered for \code{arglag} in the cross-basis formula.}

\item{knots1}{Degrees of freedom for the \code{arglag1} argument, if any.}

\item{arglag2}{Shape of relationship between lagged exposure and outcome.
Value entered for \code{arglag} in the cross-basis formula.
This setting is optional.}

\item{knots2}{Degrees of freedom for the \code{arglag2} argument, if any.
This setting is optional.}

\item{startlag}{Maximum number of lags to include in the model.}

\item{greens}{Green space values.}
}
\description{
This function determines the number of lags that should be used in the
cross-basis model. It begins with the maximum number of interest
(the startlag option). Each time a cross-basis is created, the final lag
is evaluated using unadjusted and adjusted models to determine significance
at alpha = 0.05 and alpha = 0.10. If the final lag is not significant at one
of these alphas, the model is run again with the oldest lag removed.

The output includes the settings for the final cross-basis selected,
with the final lag being the oldest with significance at alpha = 0.05 if
applicable or, if none have significance at alpha = 0.05, the oldest final
lag with significance at alpha = 0.10 is retained. If no final lags have
significance at alpha = 0.05 or 0.10, the smallest possible cross-basis is
retained, with the number of lags being 1 plus the number of degrees of
freedom. For linear models, one lag was retained.

If two cross-bases are used in the model, the same number of lags is used
for both and significance is determined if either one is significant.
}
