% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_aim2_forest.R
\name{draw_aim2_forest}
\alias{draw_aim2_forest}
\title{Draw a forest plot of air pollutant by green space}
\usage{
draw_aim2_forest(data, defect_abbr, lags, defect_name = "")
}
\arguments{
\item{data}{Dataset of model settings, including lag RRs and CIs.}

\item{defect_abbr}{Abbreviation for birth defect to be plotted, from the
dataset.}

\item{lags}{Maximum number of lags in the model.}

\item{defect_name}{Full name of the defect, for the graph title.}
}
\description{
This function reads in a table with risk ratios and confidence intervals by
lag to be plotted on a graph, then calls \code{draw_aim2_forest_plot()} to
create two ggplot objects (one for mean concentrations and one for peaks)
and uses a grob function to combine them into one figure.
}
