% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rr.R
\name{extract_rr}
\alias{extract_rr}
\title{Extract model risk ratios}
\usage{
extract_rr(model, label, data)
}
\arguments{
\item{model}{Model settings.}

\item{label}{Model label (name from a model list).}

\item{data}{Dataset.}
}
\description{
This function extracts risk ratios from two different cross-basis objects
when both are used in the same model.

At the moment, cross-prediction settings (notably centering values) are
hard-coded, but that may be changed later.

Note to self: I modified this at some point - make sure it still works when
only one cross-basis is used.
}
