% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot.R
\name{residual_plot}
\alias{residual_plot}
\title{Create a basic residual plot}
\usage{
residual_plot(bd, air, measure, model, list)
}
\arguments{
\item{bd}{String denoting birth defect}

\item{air}{String denoting air pollutant}

\item{measure}{String denoting type of pollutant measure}

\item{model}{String denoting type of cross-basis model}

\item{list}{List object containing the series of models

This function reads in a list produced by \code{cb_signif} or \code{cb_diagnostics}
and returns a residual plot with basic, but correct, labels and an overlaid
best fit line.}
}
\description{
Create a basic residual plot
}
