% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_by_buffer.R
\name{rr_by_buffer}
\alias{rr_by_buffer}
\title{Extract model risk ratios}
\usage{
rr_by_buffer(models, buffers)
}
\arguments{
\item{models}{List of settings for multiple models.}

\item{buffers}{Green space buffers.}
}
\description{
This function reads in a list of models and loops through them to extract
risk ratios and other details.
It returns a tibble with the details about each model.
}
