% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Run the models}
\usage{
run_models(settings, data, buffers = NULL)
}
\arguments{
\item{settings}{Model settings table.}

\item{data}{Dataset to model.}

\item{buffers}{Green space buffers, if relevant.

This function runs the multi-pollutant models for Aim 3.}
}
\description{
Run the models
}
