% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pred.R
\name{summarize_pred}
\alias{summarize_pred}
\title{Summarize predictor details}
\usage{
summarize_pred(pred, threshold, air, lbl)
}
\arguments{
\item{pred}{Prediction model.}

\item{threshold}{Centering value used in the prediction.}

\item{air}{Air pollutant.}

\item{lbl}{Label for the model.}
}
\description{
This function extracts parameters of the prediction model to be used for
assessment and plotting.
}
